perpdex-acc-monitor/requirements.txt
requests
python-dotenv
pyyaml
telebot
pillow  # 用于生成图片
apscheduler  # 定时任务